+++ lib/datasource/remote/api/balance_repository.dart
@@ -1,10 +1,11 @@
 import 'package:async/async.dart';
 import 'package:http/http.dart' as http;
-import 'package:seeds/datasource/remote/api/http_repo/network_repository.dart';
+import 'package:seeds/datasource/remote/api/http_repo/http_repository.dart';
 import 'package:seeds/datasource/remote/model/balance_model.dart';
 
 class BalanceRepository extends HttpRepository {
-  Future<Result<BalanceModel>> getTokenBalance(String userAccount, {required String tokenContract, required String symbol}) {
+  Future<Result<BalanceModel>> getTokenBalance(String userAccount,
+      {required String tokenContract, required String symbol}) {
     print('[http] get seeds getTokenBalance $userAccount for $symbol');
 
     final String request = '''
