+++ lib/datasource/remote/api/signup_repository.dart
@@ -8,7 +8,7 @@ import 'package:eosdart_ecc/eosdart_ecc.dart';
 import 'package:firebase_dynamic_links/firebase_dynamic_links.dart';
 import 'package:http/http.dart' as http;
 import 'package:seeds/datasource/remote/api/eos_repository.dart';
-import 'package:seeds/datasource/remote/api/http_repo/network_repository.dart';
+import 'package:seeds/datasource/remote/api/http_repo/http_repository.dart';
 
 class SignupRepository extends EosRepository with HttpRepository {
   Future<Result> unpackDynamicLink(String scannedLink) async {
